Background added and displaying in correct size
